#22 - Plug into the post-autoload-dump event to install extensions
Trying to install packages manually is not a safe method. Composer
does a lot of internal problem solving to figure out the desired
version of a package and there is no way to easily plug into this
API. The proper solution will be to queue the extension packages
and install them *after* Composer is finished downloading,
installing and creating the autoloader.